[TS-280] Refactor: my page modal refactoring
