+++ src/components/MyPage/Modal/DeleteHistoryModal.tsx
@@ -1,12 +1,9 @@
+import FooterBtn from "@/components/common/FooterBtn/FooterBtn";
 import Modal from "@/components/common/Modal/Modal";
+import { layoutStyle } from "@/components/MyPage/Modal/LogoutModal";
 
 import { useAppDispatch } from "@/api/hooks";
 import { closeModal } from "@/api/modal/modalSlice";
-
-import {
-	layoutStyle,
-	buttonBoxStyle,
-} from "@/components/MyPage/MyInfo/LogoutModal/LogoutModal.style";
 
 const DeleteHistoryModal = () => {
 	const dispatch = useAppDispatch();
@@ -16,12 +13,12 @@ const DeleteHistoryModal = () => {
 			<div css={layoutStyle}>
 				<h1>지난 히스토리를 삭제할까요?</h1>
 
-				<div css={buttonBoxStyle}>
-					<button className="cancel" onClick={() => dispatch(closeModal())}>
-						취소
-					</button>
-					<button>삭제</button>
-				</div>
+				<FooterBtn
+					text="삭제"
+					leftText="취소"
+					handleLeftBtnClick={() => dispatch(closeModal())}
+					isPositionStatic
+				/>
 			</div>
 		</Modal>
 	);
