+++ src/components/MyPage/Modal/SelectCharacterModal/SelectCharacterModal.style.ts
@@ -4,7 +4,7 @@ import { theme } from "@/styles/theme";
 
 export const layoutStyle = css`
 	width: 100%;
-	padding: 1.125rem 1.5rem;
+	padding: 1.125rem 1.5rem 5.4375rem;
 	border-radius: 15px 15px 0 0;
 	color: ${theme.color.font_black};
 	background-color: #fff;
@@ -32,24 +32,3 @@ export const getCharacterBoxStyle = (isSelect: boolean) => {
 		border: ${isSelect ? `5px solid ${theme.color.main_blue}` : "none"};
 	`;
 };
-
-export const buttonBoxStyle = css`
-	display: flex;
-	align-items: center;
-	gap: 8px;
-	margin-top: 1rem;
-
-	& > button {
-		width: 50%;
-		height: 3.4375rem;
-		border-radius: 15px;
-		color: #fff;
-		background-color: ${theme.color.main_blue};
-		${theme.font.button_big};
-
-		&.cancel {
-			background-color: ${theme.color.button_disabled};
-			color: ${theme.color.button_deactivated};
-		}
-	}
-`;
