+++ src/components/MyPage/Modal/SelectIdentityModal/SelectIdentityModal.style.ts
@@ -3,7 +3,7 @@ import { css } from "@emotion/react";
 import { theme } from "@/styles/theme";
 
 export const layoutStyle = css`
-	padding: 1.125rem 1.5rem;
+	padding: 1.125rem 1.5rem 5.4375rem;
 	border-radius: 15px 15px 0 0;
 	color: ${theme.color.font_black};
 	background-color: #fff;
