+++ src/components/common/Modal/CommonModal/SelectAgeModal.tsx
@@ -2,39 +2,35 @@ import { Dispatch, SetStateAction, useState } from "react";
 
 import CheckIcon from "@/assets/icon/ic-check.svg?react";
 
+import FooterBtn from "@/components/common/FooterBtn/FooterBtn";
 import Modal from "@/components/common/Modal/Modal";
 
 import { useAppDispatch } from "@/api/hooks";
 import { closeModal } from "@/api/modal/modalSlice";
 
-import { selectGenderData } from "@/constants/myPageConstants";
+import { selectAgeData } from "@/constants/myPageConstants";
 
-import { buttonBoxStyle } from "@/components/MyPage/MyInfo/SelectCharacterModal/SelectCharacterModal.style";
 import {
 	layoutStyle,
 	getCheckBoxLabelStyle,
-} from "@/components/MyPage/MyInfo/SelectGenderModal/SelectGenderModal.style";
+} from "@/components/common/Modal/CommonModal/SelectGenderModal/SelectGenderModal.style";
 
-const SelectGenderModal = ({
-	changeGender,
-}: {
-	changeGender: Dispatch<SetStateAction<string>>;
-}) => {
+const SelectAgeModal = ({ changeAge }: { changeAge: Dispatch<SetStateAction<string>> }) => {
 	const dispatch = useAppDispatch();
 
 	const [checkItem, setCheckItem] = useState("");
 
 	const handleCheckClick = () => {
-		changeGender(checkItem);
+		changeAge(checkItem);
 		dispatch(closeModal());
 	};
 
 	return (
 		<Modal isBottomSheet>
 			<div css={layoutStyle}>
-				<h1>성별을 선택해 주세요</h1>
+				<h1>나이대를 선택해 주세요</h1>
 				<ul>
-					{selectGenderData.map((data) => (
+					{selectAgeData.map((data) => (
 						<li key={data.text}>
 							<input type="checkbox" id={data.text} onChange={() => setCheckItem(data.text)} />
 							<label htmlFor={data.text} css={getCheckBoxLabelStyle(checkItem === data.text)}>
@@ -45,15 +41,15 @@ const SelectGenderModal = ({
 					))}
 				</ul>
 
-				<div css={buttonBoxStyle}>
-					<button className="cancel" onClick={() => dispatch(closeModal())}>
-						취소
-					</button>
-					<button onClick={handleCheckClick}>확인</button>
-				</div>
+				<FooterBtn
+					text="확인"
+					leftText="취소"
+					handleBtnClick={handleCheckClick}
+					handleLeftBtnClick={() => dispatch(closeModal())}
+				/>
 			</div>
 		</Modal>
 	);
 };
 
-export default SelectGenderModal;
+export default SelectAgeModal;
